requirements.txt
Django==4.2.7
django-extensions==3.2.3
openpyxl==3.1.2
pandas==2.1.3
python-decouple==3.8
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-import-export==3.3.1
psycopg2-binary==2.9.9
requests==2.31.0